+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- **CUMULUS-2687**
+  - Added breadcrumbs to Executions parent and children pages
+
 ### Changed
 
 - **CUMULUS-NONE**